Use sanitized domain id with Intel IOMMU
The domain id for Intel IOMMUs depend on read out hardware features and
may not be identical to the 'did' as used for host pagetables. Use as domain
id for Intel IOMMUs the one introduced with AMD IOMMU in the Space_mem
implemenation. Restrict the range depending on the count supported by the
hardware.