Fix IOTLB flush for global mode
The typo was introduced in

  4ab40e4 Flush IOMMU caches on memory revoke

and always overwrote bit 2 of the G field (granularity). With
global-granularity mode (01b) specified, the result was
Page-Selective-within-Domain Invalidation (11b), which is interpreted as
an invalid descriptor on hardware lacking this feature. The error was
apparent in REG_FSTS bit 4, which is not checked by the current
implementation and results in an endless (polling) wait for termination
of the queued invalidation.

Fixes #6